+++ examples/inverse/plot_lcmv_beamformer_volume.py
@@ -11,15 +11,9 @@ Freeview.
 #
 # License: BSD (3-clause)
 
-import numpy as np
-import matplotlib.pyplot as plt
-
 import mne
 from mne.datasets import sample
 from mne.beamformer import make_lcmv, apply_lcmv
-
-from nilearn.plotting import plot_stat_map
-from nilearn.image import index_img
 
 print(__doc__)
 
@@ -30,12 +24,16 @@ print(__doc__)
 # Data preprocessing:
 
 data_path = sample.data_path()
+subjects_dir = data_path + '/subjects'
 raw_fname = data_path + '/MEG/sample/sample_audvis_raw.fif'
 event_fname = data_path + '/MEG/sample/sample_audvis_raw-eve.fif'
 fname_fwd = data_path + '/MEG/sample/sample_audvis-meg-vol-7-fwd.fif'
 
 # Get epochs
 event_id, tmin, tmax = [1, 2], -0.2, 0.5
+
+# Read forward model
+forward = mne.read_forward_solution(fname_fwd)
 
 # Setup for reading the raw data
 raw = mne.io.read_raw_fif(raw_fname, preload=True)
@@ -71,9 +69,6 @@ noise_cov = mne.compute_covariance(epochs, tmin=tmin, tmax=0, method='shrunk')
 data_cov = mne.compute_covariance(epochs, tmin=0.04, tmax=0.15,
                                   method='shrunk')
 
-# Read forward model
-forward = mne.read_forward_solution(fname_fwd)
-
 # Compute weights of free orientation (vector) beamformer with weight
 # normalization (neural activity index, NAI). Providing a noise covariance
 # matrix enables whitening of the data and forward solution. Source orientation
@@ -92,30 +87,17 @@ stc = apply_lcmv(evoked, filters, max_ori_out='signed')
 ###############################################################################
 # Plot source space activity:
 
-# take absolute values for plotting
-stc.data[:, :] = np.abs(stc.data)
+# You can save result in stc files with:
+# stc.save('lcmv-vol')
 
-# Save result in stc files
-stc.save('lcmv-vol')
+clim = dict(kind='value', pos_lims=[0.3, 0.6, 0.9])
+stc.plot(src=forward['src'], subject='sample', subjects_dir=subjects_dir,
+         clim=clim)
 
-stc.crop(0.0, 0.2)
+###############################################################################
+# We can also visualize the activity on a "glass brain" (shown here with
+# absolute values):
 
-# Save result in a 4D nifti file
-img = stc.as_volume(forward['src'], mri_resolution=False)
-
-t1_fname = data_path + '/subjects/sample/mri/T1.mgz'
-
-# Plotting with nilearn ######################################################
-# Based on the visualization of the sensor space data (gradiometers), plot
-# activity at 88 ms
-idx = stc.time_as_index(0.088)
-plot_stat_map(index_img(img, idx), t1_fname, threshold=0.45,
-              title='LCMV (t=%.3f s.)' % stc.times[idx])
-
-# plot source time courses with the maximum peak amplitudes at 88 ms
-plt.figure()
-plt.plot(stc.times, stc.data[np.argsort(np.max(stc.data[:, idx],
-                                               axis=1))[-40:]].T)
-plt.xlabel('Time (ms)')
-plt.ylabel('LCMV value')
-plt.show()
+clim = dict(kind='value', lims=[0.3, 0.6, 0.9])
+abs(stc).plot(src=forward['src'], subject='sample', subjects_dir=subjects_dir,
+              mode='glass_brain', clim=clim)